+++ apps/time_master/src/db_inserter/inserter/facade/DatabaseInserter.hpp
@@ -1,7 +1,6 @@
+// db_inserter/inserter/facade/DatabaseInserter.hpp
 #ifndef DATABASE_INSERTER_HPP
 #define DATABASE_INSERTER_HPP
-
-// db_inserter/inserter/DatabaseInserter.h
 
 #include <string>
 #include <vector>
@@ -9,9 +8,9 @@
 #include <memory>
 #include "db_inserter/model/time_sheet_model.hpp"
 
-#include "db_inserter/inserter/_internal/DataInserter.hpp"
-#include "db_inserter/inserter/_internal/DbConnectionManager.hpp"
-#include "db_inserter/inserter/_internal/DbStatementManager.hpp"
+#include "db_inserter/inserter/pipelines/DataInserter.hpp"
+#include "db_inserter/inserter/pipelines/DbConnectionManager.hpp"
+#include "db_inserter/inserter/pipelines/DbStatementManager.hpp"
 
 
 /**