+++ apps/time_master/src/queries/shared/utils/query_utils.hpp
@@ -1,3 +1,4 @@
+// queries/shared/utils/query_utils.hpp
 #ifndef QUERY_UTILS_HPP
 #define QUERY_UTILS_HPP
 
@@ -6,7 +7,7 @@
 #include <vector>
 #include <map>
 #include <sstream>
-#include "common/common_utils.hpp" // For ProjectNode, ProjectTree
+#include "common/utils/ProjectTree.hpp" // For ProjectNode, ProjectTree" // For ProjectNode, ProjectTree
 #include "queries/shared/ReportFormat.hpp"